fair-gig-monorepo/services/profile-service/requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-dotenv
cryptography
pymongo
